+++ REACT/react-base/src/components/Header/index.js
@@ -6,7 +6,7 @@ import { useSelector } from 'react-redux';
 import { Nav } from './styled';
 
 export default function Header() {
-  const botaoClicado = useSelector(state => state.botaoClicado);
+  const botaoClicado = useSelector(state => state.example.botaoClicado);
 
   return (
     <Nav>
