+++ REACT/react-base/src/pages/Login/index.js
@@ -3,6 +3,7 @@ import { useDispatch } from 'react-redux';
 
 import { Container } from '../../styles/GlobalStyles';
 import { Title, Paragrafo } from './styled';
+import * as exampleActions from '../../store/modules/example/actions';
 
 export default function Login() {
   const dispatch = useDispatch();
@@ -10,9 +11,7 @@ export default function Login() {
   function handleClick(e) {
     e.preventDefault();
 
-    dispatch({
-      type: 'BOTAO_CLICADO',
-    });
+    dispatch(exampleActions.clicaBotao());
   }
 
   return (
