+++ REACT/react-base/src/store/modules/example/actions.js
@@ -1,0 +1,7 @@
+import * as types from '../types';
+
+export function clicaBotao() {
+  return {
+    type: types.BOTAO_CLICADO,
+  };
+}
