+++ REACT/react-base/src/store/modules/example/reducer.js
@@ -1,0 +1,17 @@
+const initialState = {
+  botaoClicado: false,
+};
+
+export default function(state = initialState, action) {
+  switch (action.type) {
+    case 'BOTAO_CLICADO': {
+      const newState = { ...state };
+      newState.botaoClicado = !newState.botaoClicado;
+      return newState;
+    }
+
+    default: {
+      return state;
+    }
+  }
+}
