+++ REACT/react-base/src/store/modules/rootReducer.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux';
+
+import exampleReducer from './example/reducer';
+
+export default combineReducers({
+  example: exampleReducer,
+});
