+++ REACT/react-base/src/store/modules/types.js
@@ -1,0 +1,1 @@
+export const BOTAO_CLICADO = 'BOTAO_CLICADO';
